+++ src/main/java/io/github/astrapi69/spring/controller/AbstractRestController.java
@@ -46,7 +46,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import io.github.astrapi69.bean.mapper.GenericModelMapper;
-import io.github.astrapi69.collections.map.MapFactory;
+import io.github.astrapi69.collection.map.MapFactory;
 import io.github.astrapi69.data.identifiable.Identifiable;
 import io.github.astrapi69.spring.service.api.GenericService;
 import io.swagger.annotations.ApiImplicitParam;